# docs/swagger.yaml
basePath: /
definitions:
...
  internal_controllers.CreateProductInput:
    properties:
      category:
        example: Electronics
        type: string
      description:
        example: High-performance laptop
        type: string
      image_url:
        example: https://example.com/image.jpg
        type: string
      name:
        example: Laptop
        type: string
      price:
        example: 999.99
        type: number
      quantity:
        example: 10
        type: integer
    required:
    - name
    - price
    - quantity
    type: object
  internal_controllers.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  internal_controllers.MessageResponse:
    properties:
      message:
        example: success
        type: string
    type: object
...
  internal_controllers.Product:
    properties:
      category:
        example: Electronics
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: High-performance laptop
        type: string
      id:
        example: 1
        type: integer
      image_url:
        example: https://example.com/image.jpg
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Laptop
        type: string
      price:
        example: 999.99
        type: number
      quantity:
        example: 10
        type: integer
      slug:
        example: laptop
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  internal_controllers.ProductListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_controllers.Product'
        type: array
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      totalPages:
        example: 10
        type: integer
    type: object
...
  internal_controllers.UpdateProductInput:
    properties:
      category:
        example: Electronics
        type: string
      description:
        example: High-performance laptop
        type: string
      image_url:
        example: https://example.com/image.jpg
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Laptop
        type: string
      price:
        example: 999.99
        type: number
      quantity:
        example: 10
        type: integer
    type: object
  internal_controllers.UploadResponse:
    properties:
      message:
        example: uploaded
        type: string
      url:
        example: /uploads/20240101120000.jpg
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ecommerce.com
    name: API Support
  description: This is a backend service for the Go e-commerce project with comprehensive
    authentication, product management, cart, orders, and payment functionality.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Commerce API
  version: "1.0"
paths:
...
  /api/admin/products:
    post:
      consumes:
      - application/json
      description: Creates a new product in the system
      parameters:
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.CreateProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_controllers.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new product (Admin only)
      tags:
      - Products
  /api/admin/products/{id}:
    delete:
      description: Soft deletes a product from the system
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a product (Admin only)
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Updates product details
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product Update Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.UpdateProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a product (Admin only)
      tags:
      - Products
...
  /api/upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a product image to local storage
      parameters:
      - description: Product Image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.UploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload product image (Local)
      tags:
      - Upload
  /api/upload/product:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a product image to AWS S3
      parameters:
      - description: Product Image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.UploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload product image (S3)
      tags:
      - Upload
...
  /products:
    get:
      description: Gets a paginated list of products with optional filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.ProductListResponse'
      summary: List all products
      tags:
      - Products
  /products/{slug}:
    get:
      description: Retrieves a single product by its slug (cached)
      parameters:
      - description: Product Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controllers.ErrorResponse'
      summary: Get product by slug
      tags:
      - Products
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"